# orders-service/openapi.yaml
openapi: 3.0.3
info:
  title: CampusEats Orders Service
  version: "1.1.0"
  description: REST API for creating and managing CampusEats orders with HTTP method, header, caching, conditional request, authorization, CORS and retry-safety support.
servers:
  - url: http://localhost:8082
    description: Local development server

  - url: https://api.campuseats.dev
    description: Production server
...
paths:
...
  /orders:

    options:
      tags:
        - Orders

      summary: Advertise supported methods
      operationId: ordersOptions

      responses:

        "204":
          description: Supported methods

          headers:

            Allow:
              description: Methods supported by the orders collection
              schema:
                type: string
                example: GET, POST, OPTIONS

            Access-Control-Allow-Origin:
              $ref: "#/components/headers/Access-Control-Allow-Origin"

            Access-Control-Allow-Methods:
              $ref: "#/components/headers/Access-Control-Allow-Methods"

            Access-Control-Allow-Headers:
              $ref: "#/components/headers/Access-Control-Allow-Headers"


    # --------------------------------------------------------
    # POST /orders
    # --------------------------------------------------------

    post:
      tags:
        - Orders

      summary: Create a new order
      operationId: createOrder

      security:
        - bearerAuth: []

      parameters:

        - $ref: "#/components/parameters/IdempotencyKey"

        - $ref: "#/components/parameters/AcceptHeader"

      requestBody:
        required: true

        content:

          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderRequest"

      responses:

        "200":
          description: Original order returned for an idempotent retry

          headers:

            Location:
              $ref: "#/components/headers/Location"

            Cache-Control:
              $ref: "#/components/headers/Cache-Control"

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"


        "201":
          description: Order created successfully

          headers:

            Location:
              description: URL of the newly created order
              schema:
                type: string
                example: /orders/1

            Cache-Control:
              $ref: "#/components/headers/Cache-Control"

            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"

            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"


        "400":
          description: Invalid request

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"


        "401":
          description: Missing or invalid Authorization header

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"


        "406":
          description: Requested response format is not supported

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"


        "422":
          description: Order cannot be fulfilled

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"


        "429":
          description: Rate limit exceeded

          headers:

            Retry-After:
              $ref: "#/components/headers/Retry-After"

            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"

            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"


        "503":
          description: Required downstream service is unavailable

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"


    # --------------------------------------------------------
    # GET /orders
    # --------------------------------------------------------

    get:
      tags:
        - Orders

      summary: List orders
      operationId: listOrders

      security:
        - bearerAuth: []

      parameters:

        - in: query
          name: status
          required: false

          schema:
            type: string

            enum:
              - placed
              - cancelled

          description: Filter orders by status.

        - $ref: "#/components/parameters/AcceptHeader"

      responses:

        "200":
          description: List of orders

          headers:

            Cache-Control:
              $ref: "#/components/headers/Cache-Control"

            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"

            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"

            Access-Control-Allow-Origin:
              $ref: "#/components/headers/Access-Control-Allow-Origin"

          content:
            application/json:

              schema:
                type: array

                items:
                  $ref: "#/components/schemas/Order"


        "400":
          description: Invalid query parameter

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"


        "401":
          description: Missing or invalid Authorization header

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"


        "406":
          description: Requested response format is not supported

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"


        "429":
          description: Rate limit exceeded

          headers:

            Retry-After:
              $ref: "#/components/headers/Retry-After"

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"


  # ==========================================================
  # INDIVIDUAL ORDER
  # ==========================================================
...
  /orders/{id}/cancellation:

    options:
      tags:
        - Order

      summary: Advertise cancellation methods
      operationId: cancellationOptions

      parameters:

        - $ref: "#/components/parameters/OrderId"

      responses:

        "204":
          description: Supported cancellation methods

          headers:

            Allow:
              description: Methods supported by cancellation resource
              schema:
                type: string
                example: POST, OPTIONS

            Access-Control-Allow-Origin:
              $ref: "#/components/headers/Access-Control-Allow-Origin"


    post:
      tags:
        - Order

      summary: Request order cancellation
      operationId: cancelOrder

      security:
        - bearerAuth: []

      parameters:

        - $ref: "#/components/parameters/OrderId"

        - $ref: "#/components/parameters/AcceptHeader"

      responses:

        "202":
          description: Cancellation accepted

          headers:

            Cache-Control:
              $ref: "#/components/headers/Cache-Control"

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancellationResult"


        "401":
          description: Missing or invalid Authorization header

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"


        "404":
          description: Order not found

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"


        "409":
          description: Order is not in a cancellable state

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"


        "406":
          description: Requested response format is not supported

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"


        "429":
          description: Rate limit exceeded

          headers:

            Retry-After:
              $ref: "#/components/headers/Retry-After"

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"


# ============================================================
# COMPONENTS
# ============================================================
components:
...
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT


  # ==========================================================
  # PARAMETERS
  # ==========================================================
  parameters:
    OrderId:
      in: path
      name: id
      required: true

      schema:
        type: integer
        minimum: 1

      description: Order identifier.
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true

      schema:
        type: string
        minLength: 1

      description: >
        Unique key used to safely retry the same order creation
        request without creating a duplicate order.
    AcceptHeader:
      in: header
      name: Accept
      required: false

      schema:
        type: string
        default: application/json

      description: >
        Requested response media type. The service supports application/json.
...
  headers:
    Location:
      description: URL of the created resource.

      schema:
        type: string
        example: /orders/1
...
    Cache-Control:
      description: Cache policy for the response.

      schema:
        type: string
        example: no-store
    Retry-After:
      description: Number of seconds before the client should retry.

      schema:
        type: integer
        example: 60
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the rate-limit window.

      schema:
        type: integer
        example: 60
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current rate-limit window.

      schema:
        type: integer
        example: 59
    Access-Control-Allow-Origin:
      description: Allowed cross-origin origin.

      schema:
        type: string
        example: "*"
    Access-Control-Allow-Methods:
      description: HTTP methods allowed for CORS.

      schema:
        type: string
        example: GET, POST, PATCH, OPTIONS
    Access-Control-Allow-Headers:
      description: Request headers allowed for CORS.

      schema:
        type: string
        example: Content-Type, Accept, Authorization, Idempotency-Key, If-Match
...
  schemas:
    PlaceOrderRequest:

      type: object

      required:
        - studentId
        - items
        - deliveryAddressId
        - paymentMethodId

      properties:

        studentId:
          type: integer
          minimum: 1


        items:
          type: array
          minItems: 1

          items:
            $ref: "#/components/schemas/OrderItemRequest"


        deliveryAddressId:
          type: integer
          minimum: 1


        paymentMethodId:
          type: integer
          minimum: 1
    OrderItemRequest:

      type: object

      required:
        - itemId
        - qty

      properties:

        itemId:
          type: integer
          minimum: 1


        qty:
          type: integer
          minimum: 1
...
    Order:

      type: object

      required:
        - orderId
        - status
        - total
        - estimatedMinutes
        - createdAt

      properties:

        orderId:
          type: integer
          minimum: 1


        status:
          type: string

          enum:
            - placed
            - cancelled


        total:
          type: integer
          minimum: 0

          description: Total amount in paise.


        estimatedMinutes:
          type: integer
          minimum: 0


        createdAt:
          type: string
          format: date-time
    CancellationResult:

      type: object

      required:
        - orderId
        - status

      properties:

        orderId:
          type: integer
          minimum: 1


        status:
          type: string

          enum:
            - cancelled
    Problem:

      type: object

      required:
        - type
        - title
        - status
        - detail

      properties:

        type:
          type: string
          example: /errors/order-not-found


        title:
          type: string
          example: Order not found


        status:
          type: integer
          example: 404


        detail:
          type: string
          example: No order 999


        errors:
          type: array

          items:
            type: array

            items:
              type: string